Consolidate Ctrl+Break handling into it's own go package.
