+++ internal/app/mmlogicapi/mmlogicapi.go
@@ -23,12 +23,11 @@ package mmlogicapi
 
 import (
 	"errors"
-	"os"
-	"os/signal"
 
-	"github.com/GoogleCloudPlatform/open-match/internal/app/mmlogicapi/apisrv"
 	"github.com/GoogleCloudPlatform/open-match/config"
+	"github.com/GoogleCloudPlatform/open-match/internal/app/mmlogicapi/apisrv"
 	"github.com/GoogleCloudPlatform/open-match/internal/metrics"
+	"github.com/GoogleCloudPlatform/open-match/internal/signal"
 	redisHelpers "github.com/GoogleCloudPlatform/open-match/internal/statestorage/redis"
 
 	log "github.com/sirupsen/logrus"
@@ -99,8 +98,7 @@ func RunApplication() {
 	}
 
 	// Exit when we see a signal
-	terminate := make(chan os.Signal, 1)
-	signal.Notify(terminate, os.Interrupt)
-	<-terminate
+	wait, _ := signal.New()
+	wait()
 	mlLog.Info("Shutting down gRPC server")
 }
