+++ internal/signal/signal.go
